Create Restful
This is the main program that can be called by the jAgent.  It will take a JSON request, convert to an object, call a subroutine based on the request, convert the response object back to JSON, and return the JSON back to jAgent.  It also includes some basic logic for security.